test(browser): cover Copy Image JavaScript completion error
